Remove Strands, make Strand 'Hammable'
Strands didn't really give me anything beyond a way to manipulate two
Strand objects. It seemed unecessary if I could have a Strand know how
to do the maniuplation. Like, if I want to add 1 + 2, I don't have an
Integers object that does the math, the Integer object knows how to do
it.

So Strands is gone and there's a new, badly-named Hammable module that
Strand includes. It defines the - method so that I can get the
difference between two Hammable objects.

This makes everything easier to follow, I think.

The - method has a terrible implementation. That's certainly something
to improve. And I could use delegation in Strand for the array methods.