Remove HammingDifference
Once I created HammingDifference, it was clearly just Hamming with a
different name. So let's ditch it and put everything back into Hamming.

Breaking down the too-long `compute` method by making it more of a
declarative builder. This way I can move some of the logic that ended up
in there down to instance methods.