First passing code for complement.
There is duplication here, both of behavior and of knowledge.

Duplication of behavior is code approaches that look the same, like the
case statements in the `find_dna` and `find_rna` methods. Those look the
same, so extraction seems beneficial. But of the 4 case options, only 2
are shared between the methods. So extraction might make things more
difficult.

Duplication of knowledge is in the two public API methods. They both
expect a string and then they are going to return its complements. They
both know how clients are going to use this code. That duplicated
knowledge seems like a good thing to dry up