Shouldn't strands build their own complement
Previous commit had RNA and DNA as a single nucleotide, but that left
them unable to deal with strands, this Complement had to handle strands.
That felt odd. So now we have RNA and DNA as something with strands and
that can break down strands into nucleotides. Complement just serves as
a thin wrapper around these classes