In which I show off my poor knowledge of biology
What we're trying to capture here is what DNA turns into when it becomes
RNA, and vice versa, right? I'm a little vague on the biology,
admittedly. This whole 'complement' thing seems like a different way of
saying Ruby's `to_` idiom. So, while keeping the current Complement API,
I'm going to rewrite the underling code to use `to_dna` and `to_rna`

Along with this, I'm reversing the responsibility for knowing
complements. Previously, RNA would check itself for a complement to one
of its own nucleotides, but I think the best way to model this is for it
to ask DNA. It's converting `to_dna` so it seems like DNA would know the
right nucleotide matches. But, again, I probably got a C- in High School Biology